Started extracting tests out via temporary design
